refactor: Move crafting and salvage into their own components
- Add RecipeCraftingProcess.svelte
- Add ComponentSalvageProcess.svelte
- Set up bi-directional binding for component salvage and recipe crafting process to share state
